// src/app/page/master/golonganDarah-form/golonganDarah-form.interface.ts
import {EventEmitter} from '@angular/core';

export interface GolonganDarah {
    kodeGolonganDarah?: string;
	namaGolonganDarah?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	label?: any;

}